Enhance packages organisation
