--- app/src/main/java/com/infinum/daggeronboarding/SuperHeroApplication.kt
@@ -1,9 +1,0 @@
-package com.infinum.daggeronboarding
-
-import android.app.Application
-
-class SuperHeroApplication : Application() {
-
-    val marvelComponent: MarvelComponent = DaggerMarvelComponent.create()
-
-}
