+++ app/src/main/kotlin/com/infinum/daggeronboarding/di/MarvelComponent.kt
@@ -1,5 +1,6 @@
-package com.infinum.daggeronboarding
+package com.infinum.daggeronboarding.di
 
+import com.infinum.daggeronboarding.ui.HeroHeadquartersActivity
 import dagger.Component
 
 @Component(
