+++ app/src/main/kotlin/com/infinum/daggeronboarding/ui/ActivityViewBindingProperty.kt
@@ -1,0 +1,11 @@
+package com.infinum.daggeronboarding.ui
+
+import android.view.LayoutInflater
+import androidx.appcompat.app.AppCompatActivity
+import androidx.viewbinding.ViewBinding
+
+inline fun <T : ViewBinding> AppCompatActivity.viewBinding(
+    crossinline bindingInflater: (LayoutInflater) -> T
+) = lazy(LazyThreadSafetyMode.NONE) {
+    bindingInflater.invoke(layoutInflater)
+}
