+++ app/src/main/res/layout/activity_hero_headquarters.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HeroHeadquartersActivity">
+    tools:context=".ui.HeroHeadquartersActivity">
 
     <TextView
         android:id="@+id/missionDescription"
